refactor(chat): rehome the store-called utils and stranded chat hooks
store actions imported real functions from components/Chat/util, putting
components above and below the store at once; resolveMessageEmoteParts,
the sharedChatBadges module and boundedSetAdd move to utils/, the
press-data types the overlay observables carry move to
store/chat/types/chatPressData (renderers re-export them), and the
chat-only useSeventvWs and useTwitchChannelPointsEventSub hooks move
under components/Chat/hooks. the channelLoad re-export shim for the 7tv
lifecycle actions is dropped now every consumer imports the real module.
no import from components/ remains anywhere under src/store
