+++ src/components/Chat/hooks/__tests__/useTwitchChannelPointsEventSub.test.tsx
@@ -1,12 +1,12 @@
 import { renderHook, waitFor } from '@testing-library/react-native';
 
+import { useTwitchChannelPointsEventSub } from '@app/components/Chat/hooks/useTwitchChannelPointsEventSub';
 import {
   createLoggedInAuthContextValue,
   createLoggedOutAuthContextValue,
   createTestUser,
 } from '@app/context/__tests__/__fixtures__/authContext.fixture';
 import { useAuthContext } from '@app/context/AuthContext';
-import { useTwitchChannelPointsEventSub } from '@app/hooks/useTwitchChannelPointsEventSub';
 import TwitchWsService from '@app/services/twitch-ws-service';
 
 jest.mock('@app/context/AuthContext', () => ({
