+++ src/utils/chat/__fixtures__/resolveMessageEmoteParts.perf.fixture.ts
