+++ src/utils/chat/__tests__/resolveMessageEmoteParts.test.ts
