+++ src/utils/chat/resolveMessageEmoteParts.ts
