+++ src/utils/chat/sharedChatBadges/getMessageBadges.ts
