+++ src/utils/chat/sharedChatBadges/sharedChatChannelBadgesCache.ts
